feat: Add Function class to ncxx library
This commit adds the `Function` class to the ncxx library. The `Function` class provides a generic way to store and invoke callable objects. It supports functions, lambdas, and member functions. The `Function` class can be used with any callable object that has a compatible signature.

The `Function` class is implemented using type erasure techniques. It uses an internal interface and implementation classes to store and invoke the callable object. The `Function` class supports copy construction, move construction, assignment, and invocation.

In addition to adding the `Function` class, this commit also includes unit tests for the new functionality in the `function.test.cpp` file.

This commit also includes some changes to other files in the ncxx library:
- Added 'ncxx/functional/function.test.cpp' file
- Added 'ncxx/functional/function.hpp' file
- Modified 'ncxx/memory/unique-ptr.hpp' file by adding a makeUnique function template
- Added 'ncxx/type-trait/is-base-of.hpp' file