Added Codecov and Codacy analysis (#35)
Adds Codecov integration and enable code coverage in the build process

Adds a new task to the code that generates an XML coverage report using Ninja. This task is included in the Codecov workflow and will be executed during the build process.

Adds a new file `codacy-security-scan.yml` which contains the configuration for the Codacy Security Scan workflow. The workflow is triggered on push to the `main` branch, pull requests targeting the `main` branch, and on a scheduled basis every day