Add pretty-printing support for arrays in GDB
This commit adds a new file, `pretty-print.py`, which contains a class called `ArrayPrinter`. This class provides pretty-printing support for arrays in GDB. It defines an iterator that iterates over the elements of the array and displays their indices and values. The `ArrayPrinter` class also implements other methods required by GDB's pretty-printing API.

Additionally, this commit adds a function called `nos_lookup_type` that checks if the type of a value is an array and returns an instance of `ArrayPrinter` if it is. This function is then appended to GDB's list of pretty printers.

These changes enable more readable and informative output when inspecting arrays in GDB.