Add CodeQL and Dependanbot workflow and Build Kernel Debug and Release
Introduce a new GitHub Actions workflow for CodeQL analysis to enhance security and code quality. The workflow triggers on pushes, pull requests to the main branch, and a scheduled weekly run. It sets up necessary dependencies, checks out the code, initializes CodeQL for C++ analysis, and performs the analysis.

Configured Dependabot to automatically check for updates to GitHub Actions on a weekly basis. This ensures that the workflow actions remain up-to-date with the latest improvements and security patches, enhancing automation reliability and security.

Added a build matrix to the CI workflow, enabling both debug and release builds without failing early. Updated the Xmake configuration and build steps for consistency, specifying the platform and toolchain. This ensures broader testing coverage and better alignment with project requirements. No significant impact on current processes but improves flexibility and robustness.